+++ src/shillelagh/backends/apsw/dialects/base.py
@@ -37,7 +37,6 @@ class APSWDialect(SQLiteDialect):
     def __init__(
         self,
         adapters: Optional[List[str]] = None,
-        adapter_args: Optional[Dict[str, Tuple[Any, ...]]] = None,
         adapter_kwargs: Optional[Dict[str, Dict[str, Any]]] = None,
         safe: bool = False,
         *args: Any,
@@ -45,7 +44,6 @@ class APSWDialect(SQLiteDialect):
     ):
         super().__init__(*args, **kwargs)
         self._adapters = adapters
-        self._adapter_args = adapter_args
         self._adapter_kwargs = adapter_kwargs
         self._safe = safe
 
@@ -56,7 +54,6 @@ class APSWDialect(SQLiteDialect):
         Tuple[
             str,
             Optional[List[str]],
-            Optional[Dict[str, Tuple[Any, ...]]],
             Optional[Dict[str, Dict[str, Any]]],
             bool,
             Optional[str],
@@ -68,7 +65,6 @@ class APSWDialect(SQLiteDialect):
             (
                 path,
                 self._adapters,
-                self._adapter_args,
                 self._adapter_kwargs,
                 self._safe,
                 self.isolation_level,
@@ -125,9 +121,6 @@ class APSWDialect(SQLiteDialect):
             raise ProgrammingError(f"Unsupported table: {table_name}")
 
         key = adapter.__name__.lower()
-        adapter_args = adapter.parse_uri(table_name) + raw_connection._adapter_args.get(
-            key,
-            (),
-        )
-        adapter_kwargs = raw_connection._adapter_kwargs.get(key, {})
-        return adapter(*adapter_args, **adapter_kwargs)  # type: ignore
+        args = adapter.parse_uri(table_name)
+        kwargs = raw_connection._adapter_kwargs.get(key, {})
+        return adapter(*args, **kwargs)  # type: ignore