+++ src/shillelagh/backends/apsw/dialects/safe.py
@@ -12,14 +12,12 @@ class APSWSafeDialect(APSWDialect):
     def __init__(
         self,
         adapters: Optional[List[str]] = None,
-        adapter_args: Optional[Dict[str, Any]] = None,
         adapter_kwargs: Optional[Dict[str, Dict[str, Any]]] = None,
         *args: Any,
         **kwargs: Any,
     ):
         super().__init__(*args, **kwargs)
         self._adapters = adapters
-        self._adapter_args = adapter_args
         self._adapter_kwargs = adapter_kwargs
         self._safe = True
 
@@ -30,7 +28,6 @@ class APSWSafeDialect(APSWDialect):
         Tuple[
             str,
             Optional[List[str]],
-            Optional[Dict[str, Tuple[Any, ...]]],
             Optional[Dict[str, Dict[str, Any]]],
             bool,
             Optional[str],
@@ -41,7 +38,6 @@ class APSWSafeDialect(APSWDialect):
             (
                 ":memory:",
                 self._adapters,
-                self._adapter_args,
                 self._adapter_kwargs,
                 True,
                 self.isolation_level,