+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Version 0.7.0 - 2021-MM-DD
 - Improved many small bugs in the type conversion system
 - Add ``sleep``, ``version``, and ``get_metadata`` functions
 - Add REPL command-line utility (``shillelagh``)
+- Removed ``adapter_args``, use only ``adapter_kwargs`` now
 
 Version 0.6.1 - 2021-06-22
 ==========================