+++ CHANGELOG.rst
@@ -2,11 +2,12 @@
 Changelog
 =========
 
-Version 0.7.0 - 2021-MM-DD
+Version 0.7.0 - 2021-07-01
 ==========================
 
 - Add support for DML to the GSheets adapter
 - GSheets dialect now return "main" as its schema
+- Schema prefix can now be used on table names
 - GSheets now supports defining a catalog of spreadsheets
 - Improved many small bugs in the type conversion system
 - Add ``sleep``, ``version``, and ``get_metadata`` functions